refactor: move isOwnPost to post rules file
